requirements.txt
numpy>=1.24.0
python-chess>=1.999
tensorflow>=2.15.0
tensorflow-metal>=1.1.0
pytest>=7.4.0
